Refactor tags controller
